Factor apart `aead` and `digest` modules
When `ring-aead` and `ring-digest` were combined all of the types were
placed in a common toplevel namespace.

This commit factors them apart into `aead` and `digest` modules (which
can potentially be gated on cargo features in follow-up commits)
