--- tests/digest.rs
@@ -1,1 +1,0 @@
-#![no_std]
